refactor: 合并时间相关枚举
