--- zeraora/constants/months.py
@@ -1,26 +1,0 @@
-__all__ = [
-    'Months',
-]
-
-from ..typeclasses import Items
-
-
-class Months(Items):
-    January = '1月'
-    February = '2月'
-    March = '3月'
-    April = '4月'
-    May = '5月'
-    June = '6月'
-    July = '7月'
-    August = '8月'
-    September = '9月'
-    October = '10月'
-    November = '11月'
-    December = '12月'
-
-    __properties__ = 'description',
-
-    @property
-    def description(self) -> str:
-        return self._description_
