+++ zeraora/email/email_manager.py
@@ -222,9 +222,7 @@ class TencentMailReceiveService:
         """
         from tqdm import tqdm
         from bs4 import BeautifulSoup
-        from zeraora.constants.timezones import TimeZones
-        from zeraora.constants.weeks import Weeks
-        from zeraora.constants.months import Months
+        from zeraora.constants.times import Months, Weeks, TimeZones
         email_amount = self.get_email_amount()
         mail_info = {
             x + 1: {"Subject": "", "From": "", "To": "", "Date": "", "Date_Detail": "", "Body": "", "Appendix": Any}
